clang: add -fhash-section-names=N option
Hash section names longer than N characters. Short section
names are unchanged. Default behaviour is to not hash
section names.

The algorithm used to hash section names is:
* SHA256
* Base64 encode
* Truncate to N chars

As an example, truncating "ThisSectionNameIsTooLong" to 16
characters would result in "ip9RNVxH27rCS+Ix". The
corresponding clang invocation would be

clang -fhash-sections-names=16 ..

The main benefit of using a hash is that other build and runtime
components can continue to be used without modification.

The main drawback of using a hash is that the process is
non-reversible.

To alleviate the main drawback problem, we supply an additional
option:

-fhashed-section-names=path

This outputs a mapping of the original section name to its hashed
value (only if the section was actually hashed). The format of the
output file is:

<section>\t<hash>

Signed-off-by: Christopher Friedt <chrisfriedt@gmail.com>